+++ app/src/launcher.cpp
@@ -10,7 +10,6 @@
 
 #include <config.h>
 #include <logger.h>
-#include <locations.h>
 #include <lockfile.h>
 
 LauncherApp::Launcher::Launcher(const QApplication& t_application)
@@ -39,6 +38,10 @@ LauncherApp::Launcher::Launcher(const QApplication& t_application)
             &m_mainWindow, &MainWindow::showErrorMessage,
             Qt::BlockingQueuedConnection);
 
+    connect(&m_interface, &Interface::selectInstallationLocationSignal,
+            &m_mainWindow, &MainWindow::selectInstallationLocation,
+            Qt::BlockingQueuedConnection);
+
     // UI --> Thread
     connect(&m_mainWindow, &MainWindow::cancel, &m_worker, &LauncherWorker::cancel);
 
@@ -54,6 +57,7 @@ void LauncherApp::Launcher::start()
 
     qInfo("Showing main window.");
     m_mainWindow.show();
+
 
     qInfo("Starting launcher worker.");
     m_worker.start();